* [jsfm] move shared utils to legacy framework
