+++ html5/frameworks/legacy/util/shared.js
