+++ html5/shared/index.js
@@ -4,25 +4,3 @@ import './console'
 import './objectAssign'
 import './objectSetPrototypeOf'
 import './arrayFrom'
-
-import {
-  extend,
-  def,
-  remove,
-  hasOwn,
-  bind,
-  toArray,
-  isObject,
-  isPlainObject
-} from './utils'
-
-export {
-  extend,
-  def,
-  remove,
-  hasOwn,
-  bind,
-  toArray,
-  isObject,
-  isPlainObject
-}
